Add files via upload
# Introduction

The cross entropy method is a policy method which works in the following way:
The agent plays several times and the algorithm selects 'elite rewards' episodes. For example the top 30 percent of the episodes played.
It then uses cross entropy to make the neural net imitate the actions  taken at a given state s during these elite episodes.